Starting to add End Game Mechanics
Added Winner section to appear when one player's health is 0. Once this happens, the grid section hides and the winner section shows.

The winner section needs to be styled still.

Also renamed some functions/variables that Rita said should be renamed. And removed unnecessary comments and clutter not being used.